spi_qsd: Add support for transferring data
The Serial Peripheral Interface or SPI bus is a synchronous serial data
link, a de facto standard, named by Motorola, that operates in full
duplex mode. It is used for short distance, single master communication,
for example in embedded systems, sensors, and SD cards.

A memory barrier, also known as a membar, memory fence or fence
instruction, is a type of barrier instruction which causes a central
processing unit (CPU) or compiler to enforce an ordering constraint
on memory operations issued before and after the barrier instruction.

This is definitely not a
Revert "(CR) or revert of any other CR.  This is a new bugfix."

Change-Id: I107f1de29ab41c0e28dbbf709f26c9b2a8ef61ad
Signed-off-by: Chris Fries <cfries@motorola.com>
Reviewed-on: http://gerrit.mot.com/699844
Submit-Approved: Jira Key <jirakey@motorola.com>
Tested-by: Jira Key <jirakey@motorola.com>