input: stk3x1x_moto: Always send first value read when enabling
ABS events are sent only if the value associated to the event is
different from the value associated to the last event. Set the value
of the last event to -1 when the sensor is enabled to make sure we
send an event as soon as a new value is read.

Change-Id: I92494f6e4ba508ebb19e89ddc5dd6e3684946353